Move multierrors into herrors package, closes #9
