+++ lib/std/herrors/doc.go
@@ -1,0 +1,10 @@
+// Copyright 2017 Marko Kungla. All rights reserved.
+// Use of this source code is governed by a Apache License 2.0
+// license that can be found in the LICENSE file.
+
+/*
+Package herrors implements functions to manipulate errors.
+*/
+package herrors
+
+// TODO(mkungla): Complete the docs
